Big extension: tries to decode the keys.
Truth be told, with this commit, I might be pushing this function
beyond what was it initially supposed to do: read key and print
characters.

Now I came up with some scheme for reading keys ... I don't know

It is good stuff, maybe just not for this program.

On branch main

modified:   007_Read_All_Keys.asm