KIND OF COOL: A new version of 008_Define_Keys
OK, first thing first, the 008_Define_Keys is a bit of a missnomer.
It did start as a program which defines User Defined Keys (UDKs as
I refer to them here and there), but it essentially proved to be
much more than that.  It is, in essence, a skeleton of a Speccy
game.

Having written that, it didn't age well.  It was introduced back in
the days when I thought that global variables (globally accessible
addresses) were a proper way to pass arguments between the subs,
and before the days I had the sub "Browse_Key_Rows" which browses
the key rows and, even more importantly, creates a unique code for
each key which was pressed.

At first, I thought I cold "modernize" the 008, but it proved to be
easier to define a new program, from the scratch, called
008_Define_Keys_Again.asm

I hope this new program will become the new skeleton for the game.
I also hope that the old 008_Define_Keys will soon be abandoned
and forgoten, and the new one be called something like:
008_Game_Skeleton

Let's see what future brings.

On branch main

new file:   008_Define_Keys_Again.asm
modified:   compile_all.sh